move M2Slot
