+++ common/src/disk.rs
@@ -430,94 +430,6 @@ impl From<ZpoolKind> for DiskVariant {
     }
 }
 
-/// Describes an M.2 slot, often in the context of writing a system image to
-/// it.
-#[derive(
-    Debug,
-    Clone,
-    Copy,
-    PartialEq,
-    Eq,
-    PartialOrd,
-    Ord,
-    Deserialize,
-    Serialize,
-    JsonSchema,
-    Diffable,
-    strum::EnumIter,
-)]
-#[cfg_attr(any(test, feature = "testing"), derive(test_strategy::Arbitrary))]
-pub enum M2Slot {
-    A,
-    B,
-}
-
-impl M2Slot {
-    /// Flip from `A` to `B` or vice versa.
-    pub fn toggled(self) -> Self {
-        match self {
-            Self::A => Self::B,
-            Self::B => Self::A,
-        }
-    }
-
-    /// Convert this slot to an MGS "firmware slot" index.
-    pub fn to_mgs_firmware_slot(self) -> u16 {
-        match self {
-            Self::A => 0,
-            Self::B => 1,
-        }
-    }
-
-    /// Convert a putative MGS "firmware slot" index to an `M2Slot`, returning
-    /// `None` if `slot` is invalid.
-    pub fn from_mgs_firmware_slot(slot: u16) -> Option<Self> {
-        match slot {
-            0 => Some(Self::A),
-            1 => Some(Self::B),
-            _ => None,
-        }
-    }
-}
-
-impl fmt::Display for M2Slot {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::A => f.write_str("A"),
-            Self::B => f.write_str("B"),
-        }
-    }
-}
-
-impl FromStr for M2Slot {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "a" | "A" => Ok(Self::A),
-            "b" | "B" => Ok(Self::B),
-            _ => Err(format!(
-                "unrecognized value {s} for M2 slot. \
-                 Must be one of `a`, `A`, `b`, or `B`",
-            )),
-        }
-    }
-}
-
-impl TryFrom<i64> for M2Slot {
-    type Error = anyhow::Error;
-
-    fn try_from(value: i64) -> Result<Self, Self::Error> {
-        match value {
-            // Gimlet should have 2 M.2 drives: drive A is assigned slot 17, and
-            // drive B is assigned slot 18.
-            17 => Ok(Self::A),
-            18 => Ok(Self::B),
-            _ => bail!("unexpected M.2 slot {value}"),
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
