+++ dev-tools/reconfigurator-sp-updater/src/main.rs
@@ -26,10 +26,10 @@ use nexus_types::deployment::PendingMgsUpdateSpDetails;
 use nexus_types::deployment::PendingMgsUpdates;
 use nexus_types::internal_api::views::MgsUpdateDriverStatus;
 use nexus_types::inventory::SpType;
-use omicron_common::disk::M2Slot;
 use omicron_repl_utils::run_repl_on_stdin;
 use qorb::resolver::Resolver;
 use qorb::resolvers::fixed::FixedResolver;
+use sled_agent_types::disk::M2Slot;
 use sled_hardware_types::BaseboardId;
 use slog::{info, o, warn};
 use std::collections::BTreeMap;
