+++ installinator-common/Cargo.toml
@@ -18,6 +18,7 @@ schemars.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 serde_with.workspace = true
+sled-agent-types.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
 oxide-update-engine.workspace = true
