+++ installinator/Cargo.toml
@@ -36,6 +36,7 @@ oxide-tokio-rt.workspace = true
 reqwest.workspace = true
 sha2.workspace = true
 serde_json.workspace = true
+sled-agent-types.workspace = true
 sled-hardware.workspace = true
 sled-hardware-types.workspace = true
 sled-storage.workspace = true
