+++ nexus/inventory/src/builder.rs
@@ -36,8 +36,8 @@ use nexus_types::inventory::TimeSync;
 use nexus_types::inventory::Zpool;
 use omicron_cockroach_metrics::CockroachMetric;
 use omicron_cockroach_metrics::PrometheusMetrics;
-use omicron_common::disk::M2Slot;
 use omicron_uuid_kinds::CollectionKind;
+use sled_agent_types::disk::M2Slot;
 use sled_agent_types::inventory::Inventory;
 use sled_hardware_types::BaseboardId;
 use std::collections::BTreeMap;
