+++ nexus/inventory/src/collector.rs
@@ -22,8 +22,8 @@ use nexus_types::inventory::RotPageWhich;
 use nexus_types::inventory::SpType;
 use omicron_cockroach_metrics::CockroachClusterAdminClient;
 use omicron_common::address::NTP_ADMIN_PORT;
-use omicron_common::disk::M2Slot;
 use omicron_uuid_kinds::OmicronZoneUuid;
+use sled_agent_types::disk::M2Slot;
 use sled_agent_types::inventory::OmicronZoneType;
 use sled_agent_types::inventory::ZoneKind;
 use slog::Logger;
