+++ nexus/mgs-updates/src/common_sp_update.rs
@@ -18,7 +18,7 @@ use nexus_types::deployment::ExpectedVersion;
 use nexus_types::deployment::PendingMgsUpdate;
 use nexus_types::deployment::PendingMgsUpdateDetails;
 use nexus_types::inventory::SpType;
-use omicron_common::disk::M2Slot;
+use sled_agent_types::disk::M2Slot;
 use std::fmt::Display;
 use std::net::SocketAddrV6;
 use std::time::Duration;
