+++ nexus/mgs-updates/src/driver_update/test_host_phase_1.rs
@@ -15,7 +15,7 @@ use gateway_messages::SpPort;
 use gateway_test_utils::setup::GatewayTestContext;
 use nexus_types::internal_api::views::UpdateAttemptStatus;
 use nexus_types::internal_api::views::UpdateCompletedHow;
-use omicron_common::disk::M2Slot;
+use sled_agent_types::disk::M2Slot;
 use sled_hardware_types::BaseboardId;
 use slog_error_chain::InlineErrorChain;
 use sp_sim::SimulatedSp;
