+++ nexus/mgs-updates/src/host_phase1_updater.rs
@@ -142,8 +142,8 @@ use gateway_client::types::SpComponentFirmwareSlot;
 use nexus_types::deployment::PendingMgsUpdate;
 use nexus_types::deployment::PendingMgsUpdateHostPhase1Details;
 use nexus_types::inventory::SpType;
-use omicron_common::disk::M2Slot;
 use sled_agent_client::Client as SledAgentClient;
+use sled_agent_types::disk::M2Slot;
 use sled_agent_types::inventory::BootPartitionContents;
 use slog::Logger;
 use slog::debug;
