+++ nexus/mgs-updates/src/test_util/updates.rs
@@ -35,8 +35,8 @@ use nexus_types::internal_api::views::InProgressUpdateStatus;
 use nexus_types::internal_api::views::MgsUpdateDriverStatus;
 use nexus_types::internal_api::views::UpdateAttemptStatus;
 use nexus_types::internal_api::views::UpdateCompletedHow;
-use omicron_common::disk::M2Slot;
 use omicron_uuid_kinds::SpUpdateUuid;
+use sled_agent_types::disk::M2Slot;
 use sled_hardware_types::BaseboardId;
 use slog::debug;
 use slog_error_chain::InlineErrorChain;
