+++ nexus/reconfigurator/blippy/src/checks.rs
@@ -26,11 +26,11 @@ use omicron_common::address::Ipv6Subnet;
 use omicron_common::address::SLED_PREFIX_LENGTH;
 use omicron_common::api::external::Generation;
 use omicron_common::disk::DatasetKind;
-use omicron_common::disk::M2Slot;
 use omicron_uuid_kinds::MupdateOverrideUuid;
 use omicron_uuid_kinds::OmicronZoneUuid;
 use omicron_uuid_kinds::SledUuid;
 use omicron_uuid_kinds::ZpoolUuid;
+use sled_agent_types::disk::M2Slot;
 use sled_agent_types::inventory::NetworkInterface;
 use sled_agent_types::inventory::ZoneKind;
 use std::collections::BTreeMap;
