+++ nexus/reconfigurator/planning/src/mgs_updates/mod.rs
@@ -27,8 +27,8 @@ use nexus_types::deployment::planning_report::BlockedMgsUpdate;
 use nexus_types::inventory::CabooseWhich;
 use nexus_types::inventory::Collection;
 use nexus_types::tuf_repo::TufRepoDescription;
-use omicron_common::disk::M2Slot;
 use omicron_uuid_kinds::SledUuid;
+use sled_agent_types::disk::M2Slot;
 use sled_hardware_types::BaseboardId;
 use slog::{error, info, warn};
 use slog_error_chain::InlineErrorChain;
