+++ nexus/types/src/deployment/planning_report.rs
@@ -17,7 +17,6 @@ use iddqd::IdOrdItem;
 use iddqd::id_upcast;
 use indent_write::fmt::IndentWriter;
 use omicron_common::api::external::Generation;
-use omicron_common::disk::M2Slot;
 use omicron_common::policy::BOUNDARY_NTP_REDUNDANCY;
 use omicron_common::policy::COCKROACHDB_REDUNDANCY;
 use omicron_common::policy::INTERNAL_DNS_REDUNDANCY;
@@ -29,6 +28,7 @@ use omicron_uuid_kinds::ZpoolUuid;
 use schemars::JsonSchema;
 use serde::Deserialize;
 use serde::Serialize;
+use sled_agent_types::disk::M2Slot;
 use sled_agent_types::inventory::ZoneKind;
 use sled_hardware_types::BaseboardId;
 use slog_error_chain::InlineErrorChain;
