+++ nexus/types/src/inventory/display.rs
@@ -16,10 +16,10 @@ use gateway_types::component::SpType;
 use iddqd::IdOrdMap;
 use indent_write::fmt::IndentWriter;
 use itertools::Itertools;
-use omicron_common::disk::M2Slot;
 use omicron_uuid_kinds::{
     DatasetUuid, OmicronZoneUuid, PhysicalDiskUuid, ZpoolUuid,
 };
+use sled_agent_types::disk::M2Slot;
 use sled_agent_types::inventory::{
     SvcEnabledNotOnlineState, SvcsEnabledNotOnline, SvcsEnabledNotOnlineResult,
     SvcsError,
