+++ sled-agent/config-reconciler/src/internal_disks.rs
@@ -20,9 +20,9 @@ use omicron_common::backoff::Backoff as _;
 use omicron_common::backoff::ExponentialBackoffBuilder;
 use omicron_common::disk::DiskIdentity;
 use omicron_common::disk::DiskVariant;
-use omicron_common::disk::M2Slot;
 use omicron_common::zpool_name::ZpoolName;
 use omicron_uuid_kinds::InternalZpoolUuid;
+use sled_agent_types::disk::M2Slot;
 use sled_hardware::PooledDiskError;
 use sled_storage::config::MountConfig;
 use sled_storage::dataset::CLUSTER_DATASET;
