+++ sp-sim/Cargo.toml
@@ -25,6 +25,7 @@ serde.workspace = true
 serde_cbor.workspace = true
 sha2.workspace = true
 sha3.workspace = true
+sled-agent-types.workspace = true
 slog.workspace = true
 slog-dtrace.workspace = true
 thiserror.workspace = true
