+++ wicketd/src/update_tracker.rs
@@ -32,12 +32,12 @@ use gateway_types::rot::RotState;
 use installinator_common::InstallinatorCompletionMetadata;
 use installinator_common::WriteOutput;
 use lpc55_areas::CMPAPage;
-use omicron_common::disk::M2Slot;
 use omicron_uuid_kinds::MupdateUuid;
 use oxide_update_engine::AbortHandle;
 use oxide_update_engine_types::events::ProgressUnits;
 use oxide_update_engine_types::spec::{EngineSpec, GenericSpec};
 use semver::Version;
+use sled_agent_types::disk::M2Slot;
 use sled_hardware_types::OxideSled;
 use slog::Logger;
 use slog::error;
